pyproject.toml
[tool.poetry]
name = "textcraft"
version = "0.0.3"
description = "Amplify llm capabilities through langchain"
authors = ["Archer <archeruuuu@gmail.com>"]
readme = "README.md"
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
dashscope = "*"
elasticsearch = "*"
fastapi = "*"
gradio = "*"
httpx = { version = "*", extras = ["socks"] }
kor = "*"
langchain = "*"
langserve = { version = "*", extras = ["all"] }
openai = "*"
pinecone-client = "*"
pydantic = "2.4.2"
pydantic-settings = "2.0.3"
pydantic_core = "2.10.1"
pymongo = "3.11.4"
python-dotenv = "*"
python-multipart = "*"
redis = "*"
rich = "*"
streamlit = "*"
sentinel = "1.0.0"
transformers = "*"
tiktoken = "*"
uvicorn = "*"
websocket-client = "*"

[tool.poetry.group.dev.dependencies]
black = "*"
isort = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'
packages = ["textcraft"]
extend-exclude = '.+/(dist|.venv|venv|build)/.+'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER"
]
skip = '''
    .tox
    __pycache__
    *.pyc
    .env
    venv*/*
    .venv/*
    reports/*
    dist/*
'''